+++ MekHQ/src/mekhq/campaign/randomEvents/prisoners/PrisonerEventManager.java
@@ -69,9 +69,9 @@ import mekhq.campaign.mission.rentals.ContractRentalType;
 import mekhq.campaign.mission.rentals.FacilityRentals;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.randomEvents.prisoners.prisonerEvents.PrisonEscapeScenario;
-import mekhq.campaign.randomEvents.randomEventSystem.RandomEventData;
-import mekhq.campaign.randomEvents.randomEventSystem.RandomEventEffectsManager;
-import mekhq.campaign.randomEvents.randomEventSystem.RandomEventResponseQuality;
+import mekhq.campaign.randomEvents.randomEventsSystem.RandomEventData;
+import mekhq.campaign.randomEvents.randomEventsSystem.RandomEventEffectsManager;
+import mekhq.campaign.randomEvents.randomEventsSystem.RandomEventResponseQuality;
 import mekhq.campaign.unit.Unit;
 import mekhq.campaign.universe.factionStanding.FactionStandings;
 import mekhq.gui.baseComponents.immersiveDialogs.ImmersiveDialogNotification;
